Touch area resizing / translation, touch click
New settings have been added to the absolute touchscreen mouse mode.

When Touch Click is set to 1, whenever the touch screen is being used, the mouse will also be activated.

There are now multiple settings for defining how the mouse moves with the touch screen.
Offset X and Y define the upper left corner bound of the mouse.
Size X and Y define how large of an area the mouse can inhabit. If the value is set to -1, it is set to the size of your monitor.

Min/Max X and Y describe the available that you can touch on the 3ds.